\name{ICsimul_prop}
\alias{ICsimul_prop}
\title{Confidence Intervals for Proportions}
\usage{
ICsimul_prop()
}
\description{
\preformatted{Nsim: Number of simulations.}
\preformatted{n: Sample size}
\preformatted{p: Probability}
\preformatted{conflevel: Confidence level (Default 0.95)}
\examples{
ICsimul_prop(Nsim=100,n=50,p=0.5,conflevel=0.95)
}
