\name{ICsimul_prop}
\alias{ICsimul_prop}
\title{Confidence Intervals for Proportions}
\usage{
ICsimul_prop()
}
\description{
ICsimul_prop()\
\item{Nsim: Number of simulations.}
\item{n: Sample size}
\item{p: Probability}
\item{conflevel: Confidence level (Default 0.95)}
\examples{
ICsimul_prop(Nsim=100,n=50,p=0.5,conflevel=0.95)
}
